reformatting of the code (stable), test routine written, introducing versioning and documentation
