+++ .gitignore
@@ -130,4 +130,4 @@ dmypy.json
 
 
 # IGNORE 
-/data
+data
