+++ nap/__init__.py
@@ -1,0 +1,9 @@
+import sys, os
+
+sys.path.append(os.path.abspath(""))
+sys.path.append(f"{os.path.abspath('')}/libs/dreem")
+
+import nap.data_wrangler
+import nap.firebase
+import nap.plot
+import nap.utils
