+++ pyproject.toml
@@ -1,0 +1,8 @@
+[build-system]
+requires = ["flit_core >=3.2,<4"]
+build-backend = "flit_core.buildapi"
+
+[project]
+name = "New Analysis Pipeline (NAP)"
+authors = [{name = "Yves Martin des Taillades", email = "yves@martin.yt"}]
+dynamic = ["version", "description"]
